requirements.txt
Flask==3.0.3
flask-qrcode==3.1.0
Pillow==10.4.0
piexif==1.1.3

